pyproject.toml
[project]
name = "learning-platform"
version = "0.1.0"
description = ""
authors = [
    { name = "DorLex", email = "dorozhinskiysanya@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "django (>=5.2.8,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "celery[amqp,redis] (>=5.5.3,<6.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "drf-spectacular (>=0.29.0,<0.30.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pre-commit (>=4.4.0,<5.0.0)",
    "ruff (>=0.14.5,<0.15.0)"
]


[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = [
    "E501", # проверка длины строки
    "I", # авто-сортировка импортов, isort
    "UP", # проверка на устаревшие сущности из модулей, pyupgrade
    "ANN001", # аннотации типов аргументов
    # аннотации возвращаемых типов для методов:
    "ANN201",
    "ANN202",
    "ANN204",
    "ANN205",
    "ANN206",
    "A", # названия переменных, "затеняющие" встроенные функции и т.д. (A001, A002 ...)
    "BLE", # слишком широкий отлов ошибки (except Exception:)
    "COM", # завершающие запятые (COM812 ...)
    "F", # неиспользуемые импорты и много чего еще (F401 ...)
    "ISC", # многострочные str
    "N", # pep8 naming (N801 ...)
    "T20" # забытый print() в коде
]
ignore = [
    "Q000", # установка двойных ковычек вместо одинарных
]

[tool.ruff.format]
quote-style = "single" # какие ковычки будет устанавливать форматтер, если может
